\name{quasiRNG-package}
\alias{quasiRNG-package}
\alias{quasiRNG}
\docType{package}
\title{
  General remarks on toolbox for pseudo and quasi random number generation
}
\description{The randtoolbox-package started in 2007 
  during an ISFA (France) working group. From then, it
  grew quickly thanks to the contribution of Diethelm Wuertz
  and Petr Savicky. It was presented in the Rmetrics 
  Meielisalp workshop in 2009. Slides can be found
  here: \url{http://www.rmetrics.org/Meielisalp2009/Presentations/Dutang.pdf}.
   
  This package has currently implementations for 
  the usual quasi RNGs.

  We recommend first users to take a look at the first
  part of the vignette. The second part is for people who
  want a deeper presentation of this topic.
} 
\details{
Package for quasi RNGs, which is a part of randtoolbox collection.
}
\seealso{
    Package pseudoRNG for pseudo random number generation. 
}
\author{
Christophe Dutang and Petr Savicky.
}
\keyword{ package }


