\name{setSeed}
\alias{setSeed}
\title{Function setSeed()}
\description{
 Function setSeed() to be used for initializaing torus.
}
\usage{
setSeed(seed)
}
\arguments{
  \item{seed}{a single value, interpreted as a positive integer for the seed. e.g.
  		append your day, your month and your year of birth.}
}
\details{
See the pdf vignette for details.
}
\value{
  \code{setSeed} sets the seed
}
\references{
Monte Carlo and Quasi-Monte Carlo Methods, Springer, pp 56--69. (available online)

}
\seealso{
    \code{\link{.Random.seed}} for what is done in R about random number generation.
}
\author{
 Christophe Dutang and Petr Savicky
}
\examples{

# (6) How to set the seed?
# all example is duplicated to ensure setSeed works

}
\keyword{distribution}

